pyproject.toml
[tool.poetry]
name = "wallpaperapp"
version = "0.1.0"
description = ""
authors = ["skorodenko <mskorodenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-guardian = "^2.4.0"
django-cors-headers = "^3.14.0"
drf-nested-forms = "^1.1.7"
django-imagekit = "^4.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
